requirements.txt
authy==2.2.3
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
idna==2.8
PyJWT==1.7.1
PySocks==1.7.0
pytz==2019.1
requests==2.21.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
twilio==6.26.3
urllib3==1.24.3
